Add code path for dealing with region size of 64-bits for multiUpdateAdd
Summary:
With T42487 it is expected that most plans will use a grain size of
64-bits. Having a code path specialised for such region sizes would
improve most of the cases of the use of scatter-add (for BERT,
GNNs, ....) we use in practice.

A code path is used rather than a specialisation because it costs
only 2 cycles for each call.

Ref T46649

Test Plan: New tests added

Reviewers: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T46649

Differential Revision: https://phabricator.sourcevertex.net/D52204