Allow partial types to be specified for multiUpdateAdd
Summary:
Keep temporary buffers for the base tensor when
partial type has higher dynamic range than the
data type. The result is cast before the
update to the base tensor. This replaces the
behaviour of using a higher dynamic range only
when look-up is split. This make it consistent
in all types of paritions.

The default partial type has the same type as the
type of the base tensor.

Ref T46718

Test Plan: Added new tests

Reviewers: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T46718

Differential Revision: https://phabricator.sourcevertex.net/D52405