Reduce vertex state size for strided reduce
Summary:
The outer stride is redundant in the common use case where numOuterStridesM1 = 0 .  This also creates a 16 bit gap.  It can be omitted and the same code executed saving a worthwhile percentage of the vertex state when numOuterStridesM1 = 0.

Ref T46026

Test Plan: CI, added to vertex test, and used when existing tests are appropriate.

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, claudiop, oleksiik

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, claudiop

Maniphest Tasks: T46026

Differential Revision: https://phabricator.sourcevertex.net/D52476