Fused expression vertices for a single operation involving scalars and constants
Summary:
Create fused vertices even where there is a single operation,
BUT - where all inputs are scalar
AND (One of a set of specific constants is used OR there are 2 Tensor operands, with a few specific operations)

The result is smaller vertex state (compared to a broadcast op vertex) as the length field is avoided, and constants can be embedded in the code.  The above constraints are  to avoid creating 100s of specific vertices.

This is very beneficial for loop counts

Fixes T46984

Test Plan: Added specific test to verify number of map vertices for particular expressions

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm, oleksiik

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik

Maniphest Tasks: T46984

Differential Revision: https://phabricator.sourcevertex.net/D52561