Reinstate MultiVertex for elementwise, Transpose
Summary:
All the Supervisor vertices for BinaryOp, BroadcastScalar, VectorInner,
VectorOuter, Cast and Transpose are replaced with MultiVertex.
ElementwiseStub.S is removed.
The name 'Supervisor' is systematically expunged from functions, variable
names and comments.
Also the names of 2D cast and transpose vertices is modified to use uppercase
"2D" instead of lowercase "2d" for consistency with the other vertices.

This is just a rebase of D46542 (MultiVertex for element-wise and transpose
vertices) which had been reverted because the slight increase in memory that it
caused in the control code was generating out-of memory in some Tensorflow
tests (see T40918).
Now T42414/D53997 has been landed. That will keep the size of the control code
exactly constant before and after this change, as the size of the supervisors
stubs will be the same.
The one exception to the above is that if a tile contains anyone of the 'scaled
add' vertices plus any of the vertices modified here, the total size will
increase by one word (as we no longer do 'bri __popops_SupervisorSyncAndReturn'
in ScaledAddSupervisor_fp.S)

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, timp

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, timp

Maniphest Tasks: T42837

Differential Revision: https://phabricator.sourcevertex.net/D48593