Some use of inline assembler for Unary ops
Summary:
Benefit is that we can control bundling within the inner loop for supervisor vertices.
It is important to ensure there is no strided overread.  In the worker case there would be no
stride, so no strided overread, and the loop is manually unrolled in C so no benefit to
using inline assembler

Fixes T47963

Test Plan: CI, manually check and update cycle counts for speed verification of the ops in question

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, claudiop, oleksiik, samk, ranjayc

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, claudiop

Subscribers: tomm, oleksiik, samk

Maniphest Tasks: T47963

Differential Revision: https://phabricator.sourcevertex.net/D54489