Try and allow op grainsize to be met on all tiles in intermediate -> intermediate reduction stages.
Summary:
Ref T48623.

Previously we were looking at the tiles on which partials resided for each output region and splitting the output dimension based on this to be reduced on tiles.

This meant that if the mapping was such that after the initial reduction, the size of an output region was not a multiple of grain size, we would respect that, and the resulting reduction in the intermediate -> intermediate stage would be inefficient due to not meeting grainsize for the op.

We now still respect the input tile mapping largely, and walk the output regions and merge them where the boundary is not a multiple of grain size. The result in terms of reduction is that each partial from the source tiles for the regions either side of this boundary will need to be gathered together. As those partials would have had to be exchanged between tiles anyway to be reduced (because this is an intermediate reduction this is always the case), this should be more or less free in terms of exchange code, and should at least be no worse than the current implementation except in some possible exceptional cases where one of the partials for an output region already happened to be on the tile it would be reduced on. I think this is an outlier case given we do not explicitly try and map reductions in this stage to the tiles on which the inputs resided.

This recovers the last of the performance lost in regressions in T48623 (and actually improves performance a bit).

Reviewers: timp, godfrey.da.costa, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: timp, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T48623

Differential Revision: https://phabricator.sourcevertex.net/D55428