Regroup input tensor to multi-update add with static indices
Summary:
Ideally we want to regroup both the base and input tensors
such that they have a grouping in the non-sliced dimension.
But the regrouping of the base tensor could be very costly
as it has be to done twice. Hence only attempt to regroup
the input tensor.

For a GNN sharded over 4 IPUs improves performance by
50% and saves a significant amount of vertex state as
better reduction specialisations are used with less
copying.

Ref T50474

Reviewers: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T50474

Differential Revision: https://phabricator.sourcevertex.net/D55980