Implement casting with inline assembler where needed, C++ otherwise
Summary:
Implement only the performance critical cases:
float -> half,
half -> float
char -> half
plus half-> char for symmetry
Using a full loop in inline assembler.

The remaining cases of float/half to/from 8bit types use some inline assembler to maintain the rounding and operation that was made in assembler previously but no fully optimised loop.  This will produce a regression but less assembler to maintain.

For the 1D (Multivertex) cases the regular strided method of work division is used rather than the previous method that was specific to casting vertices.

Data sizes are limited to that supported by work division and the repeat count.  Every vertex is made to deal with 4 elements per loop to make this simpler and more consistent

Ref T49866

Test Plan: CI, extend range of inputs for char casting to match the type

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, claudiop, oleksiik, ranjayc

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, claudiop

Subscribers: ranjayc

Maniphest Tasks: T49866

Differential Revision: https://phabricator.sourcevertex.net/D55475