Remove unncessary clamping of input data in erf calculation
Summary:
The clamping is not necessary because the polynomial
is a function of a value which is always less than 1
and exp(-x^2) is guaranteed to give a valid value even
for +/-Inf. also reorder the rational function such that
only an add is required The expression is made equivalent
by folding in the scale in the coefficients. This gives a
very small degradation in relative accuracy for float
but none for half.

Saves a cycle/half

Ref T39314

Reviewers: Eddy, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: Eddy, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T39314

Differential Revision: https://phabricator.sourcevertex.net/D56503