Change ConsistentExecutableTest to process-consistency
Summary:
The ConsistentExecutableTest checked that creating the same program
tree twice in the same process created bitwise-identical serialised
executables.

This is problematic for memory initialised from global state,
such as debug ids which are constructed by incrementing a static
unsigned integer at program creation time. As memory like this will
carry over state from the first program tree into the second, and
so when serialised produce different executables.

This patch relaxes the test slightly so that each executable is created
in a separate process and so starts with the same global state.

Test Plan: CI.

Reviewers: samk, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: samk, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Subscribers: hectorm

Maniphest Tasks: T46515

Differential Revision: https://phabricator.sourcevertex.net/D56989