Update group size to balance work amongst tiles for transpositions
Summary:
Use factors of the grouping to decide on a new grouping.
The target number of transpositions is set conservativelty
to the number of tiles per IPU as exchange code memory
increases with an using finer grained grouping.

Use an idealised cost model to be eventually replaced by
real estimates.

Ref T28136

Reviewers: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T28136

Differential Revision: https://phabricator.sourcevertex.net/D57110