Create a GELU_ERF unary operation in popops and use library function for float error function(erf)
Summary:
Replace the approximation used for error function (erf) in float
with the library function that has much higher absolute and
relative accurarcy. The addition of GELU_ERF unary operation
in popops allows usage of a faster lower accuracy fused operation
when a GELU with ERF is required that is more accurate than
a tanh or sigmoid approximation.

Ref T49879

Reviewers: Eddy, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: Eddy, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T49879

Differential Revision: https://phabricator.sourcevertex.net/D57228