Use 3rd order polynomial approximation for GELU-ERF in fp16
Summary:
This replaces the 5th order polynomil aproximation with
the operations in half except for the calculation of the
error function. The error performance is almost the same
because where the erf gives the largest absolute error,
erf(x/sqrt(2)) when added with 1 ~= 1 in fp16. The product
with input x further pushes the result closes to denorm values
for very small x.

Ref T49879

Reviewers: Eddy, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: Eddy, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T49879

Differential Revision: https://phabricator.sourcevertex.net/D58125