Use the quarter type provided by the compiler in Cast codelets
Summary:
Was using a locally created quarter type, which now conflicts.

The implementation with the local one allowed for a quarter to be constructed using an unsigned type.  This was just clearer for the inline assembler, knowing that the register assigned would be a $m register and that no extra code would be added to make it be treated as 8-bit even though it was contained in a 32 bit register.  However now declaring a quarter appears to function correctly.

Ref T49866

Test Plan: CI unchanged but the fact that this didn't fail implies that tests are not active.  Probably this is Ok for the moment as it makes for fewer failures until enough pieces of "quarter" support are complete for the tests to always pass.

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, ranjayc, samk

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, ranjayc

Subscribers: thomasp

Maniphest Tasks: T49866

Differential Revision: https://phabricator.sourcevertex.net/D58501