Fix to the pointer stride in the C++ version of strided reduce
Summary:
A quick fix to the issue described in T53559 (Explanation in the task)

The pointer calculated was outside the range of the data in the edge case produced in the task.  Although the pointer is never actually used a segfault was the result.

Ref T53559

Test Plan:
This is why it's a "quick fix".
The change doesn't break existing tests.
The reduce_op tool doesn't produce complex layouts that stimulate this to happen
The codelet test relies on a nice regular array on which to call a reduce function to find the reference value
So - adding proper testing is not trivial.

A quick manual test was carried out by modifying the use of the outerStride parameter in the codelet test to re-create the test case and a segfault.
With the change there was no segfault.  But as the test comparison result was is wrong the result is not verified by the test tool.  Looks correct on manual check for the problem case

Reviewers: samk, tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: samk, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T53559

Differential Revision: https://phabricator.sourcevertex.net/D58666