Add temporary reduction memory to the memory estimate in the CTC planner
Summary:
The gradients could be assigned to multiple tiles if labels are split
across tiles. As the reduction is guaranteed to be across that
dimension there cannot be a on-tile reduction. This implies that there
cannot be an in->intermediate stage. This means that a copy of the
gradients must be done via exchange. This extra copy must be included
in the memory estimate.

Ref T53693

Test Plan: Inspected plan to check for extra memory estimate field added

Reviewers: timp, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: timp, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T53693

Differential Revision: https://phabricator.sourcevertex.net/D58720