Reduce poplibs CI build test time
Summary:
Most ctc_beam_search tests are run on the IpuModel. A group of them is also
run on Sim and on the Hw.
This simulator set takes 10% of the total time of poplibs tests when running
on a local server or on the non-Hw non-Debug CI builds.
We remove this simulator run to speed up the CI builds.
These test will still be run (on Hw) in "ci_ubuntu_18_04_hw_tyan_mk2_4ipu_node"
builds.

Reviewers: timp, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: timp, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T51473

Differential Revision: https://phabricator.sourcevertex.net/D59061