popsparse matrix and convolution ops: Restrict partials type to those allowed and warn
Summary:
Where the partials type specified in options is smaller than the in/out type it is revised to the in/out type and a warning is produced.

Similarly a partials type of HALF isn't supported for elementwise sparse operations.  Again, it is revised to FLOAT and a warning produced

Ref T33267

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik, tomm

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik

Maniphest Tasks: T33267

Differential Revision: https://phabricator.sourcevertex.net/D59218