Use large tensor not a scalar to fill using copy for compile speed improvement
Summary:
Don't broadcast a scalar when filling a tensor.  Because when the destination tensor is large this increases compile time.  Instead create and map a tensor of the same size as the destination and use that.  Compile time is improved and poplar will resolve the operation to memset vertices - so the large tensor isn't actually stored on the target machine

Fixes T52168 (as fas as I can tell with local checks)

Test Plan: Checked compilation speed of full model in question and smaller reproducer (simply fill a large tensor either by broadcasting or by creating a large tensor, or using the tensors that are filled - serialised from the large model)

Reviewers: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Subscribers: vladimirm

Maniphest Tasks: T52168

Differential Revision: https://phabricator.sourcevertex.net/D59489