Use single region only if a single index is given to the planner
Summary:
Use the single region optimisation only if a single offset
is given when planning MultiSlice/MultiUpdate. This guarantees
that the optimisation is not used when statistically the
average offset per tile is 1. Also update edge alignment
to 4 bytes to guarantee that work can be split.

Ref T54537

Test Plan: Verified plans

Reviewers: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T54537

Differential Revision: https://phabricator.sourcevertex.net/D59710