Improve relative accuracy of sine and cosine functions for half type
Summary:
The relative accuracy is improved to within 1ulp for both
sine and cosine functions at all points in the half domain
except for a single point for sine and 6 points for
cosine. Cosine could use the sine implementation but the
transformation for doing so requires half->float
conversion. The worst case relative error is 1.6ulp for since
and 8 for cosine at 34571*(pi/2). All the 6 points where error
is greater than 1ulp is at a some large factor of pi/2
where cosine is close to 0.

Ref T5017

Reviewers: Eddy, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: Eddy, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T5017

Differential Revision: https://phabricator.sourcevertex.net/D60426