AMP vertices, test and planner support for fp8 inputs, weights, half partials
Summary:
Added `convPartial1x1Out` and `convPartialnx1` vertices with input, weight type = quarter and partials type = half.
Supported by Ipu21
The implementation is in C++ with inline assembler for the actual processing.  This is probably not optimal enough but is functionally correct and more maintainable.
The planner is able to target the new vertices.  Many of the changes in this area are to support change `bool floatActivations` (allowing actsType=float or half) to become `Type actsType` (as we need float, half or quarter)
Test support is added for the new input types, by casting from type half to quarter on the IPU.  The format and scale can be specified but at the moment full poplar support for this isn't available so a single hardcoded format functions correctly.

Details to complete, but possibly having to wait until poplar is updated:
Cycle estimates are a bit of a placeholder - code is in C for now but may change.
Use proper poplar quarter metadata

Ref T49867

Test Plan: 1000's conv random tests pass locally

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, mateuszk, krzysztofk, oleksiik, ranjayc, tomm

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik

Maniphest Tasks: T49867

Differential Revision: https://phabricator.sourcevertex.net/D58512