Support grouped multiUpdate/multiSlice/multiUpdateOp
Summary:
Groups are defined as independent equal sized operations
that are run in parallel. Functions to create sliceable,
slice and index tensor have a grouped vaarint.

Unplanned group operations are not supported.

Non-grouped operations are the same as a grouped operation
with that group size of 1.

Ref T41714

Test Plan: Add new test executable and tests using it.

Reviewers: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T41714

Differential Revision: https://phabricator.sourcevertex.net/D60174