Enable SLIC vertices in the planner with placeholder cycle estimates
Summary:
Already landed SLIC vertices for type quarter, half with stride 1,2.  This allows them to be targeted in the planner, and provides a structure so the cycle estimates can be modified independently of the float and half vertices.  The values in the cycle estimates are a placeholder - as the IpuModel version of the vertices is not yet ready and so checking cycle estimate accuracy will be tedious

Ref T49867

Test Plan: Verified no conv random errors and that SLIC vertices were chosen in some cases

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik, tomm, ranjayc

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik

Maniphest Tasks: T49867

Differential Revision: https://phabricator.sourcevertex.net/D61134