Treat fill of data-type Quarter as unsigned char, limit it so that the fill value =0 only
Summary:
This way no vertices are needed, and when used the metadata pointer will not be included in the
vertex state.

The interpretation of non zero values is ambiguous as it relies on the metadata and could/should be a runtime cast.  However this change serves the common need of value =0

Ref T47304

Reviewers: tomm, ranjayc, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T47304

Differential Revision: https://phabricator.sourcevertex.net/D61582