Dynamic slice for type quarter using unsigned char vertices, and preserving metadata
Summary:
This is reverting a lot of D59774, as we no longer use specific vertices of type `quarter`.  `unsigned char` is used instead and so no metadata is attached to the vertex, which doesn't need it

Relies on  D60550 for metadata support.

Most effort goes into ensuring that metadata is correctly returned

Ref T49869

Test Plan: Explicit checking of metadata values added

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, ranjayc, tomm, oleksiik

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Maniphest Tasks: T49869

Differential Revision: https://phabricator.sourcevertex.net/D61455