Set a default available memory proportion for slice planning
Summary:
If a memory proportion is not specified then use a default
value of 0.6. The value is chosen arbitrarily and is expected
that planning is done for the operation with the option
set. The default value also matches the memory proportion
used for unplanned operations.

Ref T56808

Test Plan: Verified that embedding_layer selects the correct proportion

Reviewers: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T56808

Differential Revision: https://phabricator.sourcevertex.net/D61659