Use unsigned char transpose vertices for data only from quarter tensors, minor fixup to convolution vertices
Summary:
2 small areas of change -

Use the polar vertices for transpose with an unsigned char type to carry out transpose of the data of tensors with type quarter.  This avoids attaching the metadata edge which is not necessary.

In the course of testing, there is no simple independent test of rearrangements in poplibs, so testing results in using convolution - revealing IpuModel specific bugs in the codelets which are simple to fix

Ref T54133

Test Plan: 500 each of  IpuModel and Sim tests for conv_random specifically with data type quarter

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml, ranjayc

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Maniphest Tasks: T54133

Differential Revision: https://phabricator.sourcevertex.net/D61857