Preserve the previous state in LSTM where a batch entry has stopped processing
Summary:
When running LSTM with a batch of sequences of different lengths we want to see the final state for each sequence.  Before this change the final state for sequences with a length shorter than the maximum length would return zero as the final state.  The change is similar to that done in GRU

Fixes T55480

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, ranjayc, yanislavd

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, ranjayc, yanislavd

Subscribers: godfrey.da.costa, yanislavd

Maniphest Tasks: T55480

Differential Revision: https://phabricator.sourcevertex.net/D60477