Calculations use unpadded unsliced dimension in slice planner
Summary:
The wrong assumption of base tensor being padded to a
grain size in the slice planner resulted in planner
using very high estimates for the number of base
elements. This resulted in the planner not meeting
the "plan for memory" when it actually should.

Use the existing graining for splitting work but
account for the actual elements in a grain without
padding.

This is a temporary solution until it is decided what
approach we want to take as part of T56837.

Ref T56837

Reviewers: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T56837

Differential Revision: https://phabricator.sourcevertex.net/D61806