Add option to disable stochastic rounding in AMP vertices
Summary:
This allows disabling of stochastic rounding(SR) only for
AMP vertices rather than full convolutions. A template
parameter is added to the vertex to minimise memory
overhead. C++ vertices are not modified as we do not
target AMP with them.

Tests added do not check if SR is added but only code
paths are assembler specialisations compile and link
properly.

Ref T58603

Reviewers: timp, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik

Subscribers: blazejb

Maniphest Tasks: T58603

Differential Revision: https://phabricator.sourcevertex.net/D63381