Associate metadata in convolution and dynamic slice library so that concat can not force identical metadata
Summary:
Rework poplibs functions that deal with quarter type to conform with intended rules for metadata.

Rules:
1. A variable or constant with type quarter must be created referencing a metadata tensor that was already created
2. We cannot re-associate new metadata with a tensor once created.
3. We can however create a tensor, and associate it with the metadata of another tensor
4.  When we concatenate all tensors must have the same metadata
5. Copy functions like transpose and dynamic slice make a new metadata tensor to go with the destination tensor and copy into it
6. Dynamic update has no such copy, assumes the metadata of the whole is the same as that of the slice written.  This could become a runtime check.

Implications:
a. Where we provide a poplibs function that creates a tensor, it will internally create new metadata which the caller is expected to write to.  (Cases like createWeights, createInput,  createSliceableTensor)
b. Where we create a destination tensor (createSliceTensor, createGroupedSliceTensor) a new metadata is produced internally and remains unwritten
c. Where we write to a tensor produced in case b. a copy program is included to write to the destination slice tensor (dynamicSliceWithOutput)
d. Some functions are a combination of b,c where a destination tensor is created and written to - so a copy program is added for the metadata (multiSlice, groupedMultiSlice)
e. Update doesn't add a copy

The convolution code is structured so that we could fairly easily choose to add overloads to pass a metaddata tensor in.  Where this acts as a `conditional` it is a pointer to a Tensor, using `nullptr` to indicate no metadata.  This is because `addVariable` uses `nullptr` to signify no metadata.

The dynamic slice code is not structured to allow us to add overloads to pass a metadata tensor in.  This is because to do so would mean changing an external header file.

Ref T47304

Test Plan: CI for non-quarter types.  Locally checked using quarter type.

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, ranjayc, tomm

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Subscribers: godfrey.da.costa, georgep, ranjayc

Maniphest Tasks: T47304

Differential Revision: https://phabricator.sourcevertex.net/D62090