Add scale values parameters to ScaledAddCodeletsTest
Summary:
Optional command line parameters are added to the ScaledAddCodeletsTest
executables, for specifying the values for the scaleA/scaleB fields of
the codelets.
If not specified, random values will be selected, as before.
The values used will also be printed (together with the tolerance, if
used in a specific codelet)

Test Plan: Run poplibs regression tests

Reviewers: timp, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: timp, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T39288

Differential Revision: https://phabricator.sourcevertex.net/D64660