Convolution for quarter data using cast and existing hmac,vmac vertices
Summary:
Implement convolutions with quarter data using a cast followed by existing hmac, vmac vertices with an input type of half.
Cast happens on the tile that the vertex is implemented on (and this is modelled).

Ref T49867

Test Plan:
CI for Ipu21 not yet in place.
Locally verified that plans using hmac, vmac are produced and provide the correct result.

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik, tomm

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik

Maniphest Tasks: T49867

Differential Revision: https://phabricator.sourcevertex.net/D63300