Re-enable RptLoopCycleTest_binary_float_half
Summary:
This was disabled as it failed at one point .  We believe that was just due to compile/poplar revision mismatch and it sorted itself out.

This diff just re-enables the test - probably would be just as good to revert the change that disabled the test, but at least if we land this diff there is a record that the test did pass before landing it, and what happened is more visible.

Fixes T60270

Test Plan: CI

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Maniphest Tasks: T60270

Differential Revision: https://phabricator.sourcevertex.net/D64894