Using poplar cast functions for upload/download in tests, using poplar's metadata type with Format and packing
Summary:
This relies on D63002.

Using new upload and download functions in tests to write quarter data (instead of casting from half on the IPU) and using that to write the metadata instead of a copy program.
Remove some cases where we would cast twice on the IPU to do a test which is now not necessary.
Use the QuarterMetadata and QuarterMetadata::Format types from poplar instead of the poplibs ones, along with packing function

Not Done : Where `graph.createHostWrite` and `graph.createHostRead` are used to copy tensors, tests are left as they were.  Future poplar updates should allow these tests to more elegantly create quarter type data

Ref T57103

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik

Maniphest Tasks: T57103

Differential Revision: https://phabricator.sourcevertex.net/D64130