Allow (and don't error on) expression optimisations with constants in Binary ops
Summary:
The expression optimisor would pick up expressions of the form:
_1 + (Const(1.0f) + Const(0.0f))
Where 1.0f + 0.0f should optimise into just Const(1.0f).
However a badly formed expression would be substituted at this point if the lhs (The Const1.0f in the example) was a constant, not a Tensor.
The same would apply to subtract, multiply and shl, shr where the operation would be optimised away, yet the LHS was a constant.

Fixes T57386

Test Plan:
Specific tests added
Verified with the specific test case in the task

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm, oleksiik

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik

Maniphest Tasks: T57386

Differential Revision: https://phabricator.sourcevertex.net/D65734