Removing supervisor code stubs for half data type scaled add vertices
Summary:
Removed any need for supervisor pre-processing of vertex state in scaledAdd vertices.  Some of this was already done, but this change modifies the cases where tolerance and a float scale are considered and we decide if a half scale is accurate enough

In A later diff to avoid this getting out of hand:
Make them all multivertex.
This means making the integer ones a multi vertex which will be C++ and faster than the supervisor version that exists
Maybe use the CHECK_ACCURACY_WHEN_CAST_ASM vertex in the 2D ones (intentionally not touched so far)
Consider tidying register allocations and definitions
Remove unused "isConst" template parameter

Ref T46402

Test Plan: CI.  Further verification locally that fast (float scale treated as half) cases were taken when they should be

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, claudiop, oleksiik

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, claudiop

Maniphest Tasks: T46402

Differential Revision: https://phabricator.sourcevertex.net/D65787