Remove dependency popsolver on poplibs
Summary:
Removes trace level logging of constraint types. I don't think this has been useful in optimising the planner.

This is so that popsolver can be moved to a separate library that can be used by poplar

Done ahead of moving to a new repo so the changes can be easily reviewed

Ref T55607

Test Plan: CI passes

Reviewers: mattl, tomm, samk, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: mattl, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T55607

Differential Revision: https://phabricator.sourcevertex.net/D66676