Add zeroInfinity option to CTCLoss
Summary: This option prevents the loss from going to infinity when a batch of data is too short to be aligned with the corresponding target sequence. It achieve this by masking with zeros the values of the loss and the gradients corresponding to the problematic input batches.

Test Plan:
`./test.sh poplibs`
Added tests for the new option

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, timp, godfrey.da.costa

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, godfrey.da.costa

Subscribers: godfrey.da.costa, timp

Maniphest Tasks: T61161

Differential Revision: https://phabricator.sourcevertex.net/D66459