Handle larger than machine strides in convolution C++ vertices
Summary:
Use 21-bit strides so that packed strides fit in a 64-bit
vertex field. The stride unpacking function always takes
64-bit packed stride even for 10-bit strides. But this is
not seen as a problem because the function to unpack
strides should be called for C++ vertices where strides
are 21-bit. Otherwise an overload to unpack 32-bit strides
could be added.

QUARTER type is not handled properly primarily because
unlimited version is not properly supported for that
type.

Ref T62097

Reviewers: timp, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: timp, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T62097

Differential Revision: https://phabricator.sourcevertex.net/D66884